+++ main.py
@@ -40,22 +40,6 @@ from roxbot import guild_settings as gs
 # REMEMBER TO UNCOMMENT THE GSS LINE, ROXIE
 # DO NOT UNCOMMENT GSS IF YOU ARE NOT ROXIE
 
-cogs = [
-	"roxbot.cogs.admin",
-	"roxbot.cogs.customcommands",
-	"roxbot.cogs.fun",
-	"roxbot.cogs.image",
-	"roxbot.cogs.joinleave",
-	"roxbot.cogs.nsfw",
-	"roxbot.cogs.reddit",
-	"roxbot.cogs.selfassign",
-	"roxbot.cogs.trivia",
-	"roxbot.cogs.twitch",
-	"roxbot.cogs.util",
-	"roxbot.cogs.voice",
-	#"roxbot.cogs.gss"
-]
-
 
 # Sets up Logging that discord.py does on its own
 logger = logging.getLogger('discord')
@@ -81,9 +65,6 @@ async def on_ready():
 	bot.load_extension("roxbot.system")
 	print("system.py Loaded")
 
-	bot.load_extension("roxbot.settings.settings")
-	print("settings.py Loaded")
-
 	bot.load_extension("roxbot.err_handle")
 	print("err_handle.py Loaded")
 
@@ -96,7 +77,7 @@ async def on_ready():
 
 	# Load Extension Cogs
 	print("Cogs Loaded:")
-	for cog in cogs:
+	for cog in roxbot.cogs:
 		try:
 			bot.load_extension(cog)
 			print(cog.split(".")[2])